backend/beneficiarios/serializers.py
from rest_framework import serializers
from .models import Beneficiario

class BeneficiarioSerializer(serializers.ModelSerializer):
    class Meta:
        model = Beneficiario
        fields = '__all__'
