backend/requirements.txt
django-filter

# Django Core
Django==5.0.7
djangorestframework==3.15.2
django-cors-headers==4.3.1
django-filter==23.5

# Database MySQL - usando PyMySQL (más compatible con Windows)
PyMySQL==1.1.1
mysqlclient==2.2.7


# Authentication & Authorization  
djangorestframework-simplejwt==5.3.0

# File Managementssdf
Pillow==10.4.0

# API Documentation
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.12.1

# Data Processing & Export
openpyxl==3.1.5
reportlab==4.2.2
# pandas==2.2.2  # Comentado temporalmente por problemas con Python 3.13

# Environment Management
python-decouple==3.8

# Celery for Background Tasks (opcional)
celery==5.3.4
redis==5.0.7
django-celery-beat==2.6.0

# Utilities
python-dateutil==2.9.0
requests==2.32.3

# Development Tools
django-extensions==3.2.3

# Testing
pytest-django==4.8.0
factory-boy==3.3.0
coverage==7.6.0

# Development Only
django-debug-toolbar==4.4.6
black==24.4.2
flake8==7.1.0

# Production
gunicorn==22.0.0
whitenoise==6.7.0